Split accuracy test case for reduction operators
