+++ tests/test_index_select.py
@@ -1,0 +1,41 @@
+import math
+import random
+import time
+
+import pytest
+import torch
+
+import flag_gems
+
+from . import accuracy_utils as utils
+from . import conftest as cfg
+
+if cfg.QUICK_MODE:
+    DIM_LIST = [1]
+    FLOAT_DTYPES = [torch.float32]
+else:
+    DIM_LIST = [0, 1]
+    FLOAT_DTYPES = utils.FLOAT_DTYPES
+
+random.seed(time.time() // 100)
+
+
+@pytest.mark.index_select
+@pytest.mark.parametrize("shape", utils.REDUCTION_SHAPES)
+@pytest.mark.parametrize("dim", DIM_LIST)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_index_select(shape, dim, dtype):
+    inp = torch.randn(shape, dtype=dtype, device=flag_gems.device)
+    index_size = inp.size(dim)
+
+    index = torch.randint(
+        0, index_size, [math.floor(index_size * 0.8)], device=flag_gems.device
+    )
+
+    ref_inp = utils.to_reference(inp)
+    ref_index = utils.to_reference(index)
+    ref_out = torch.index_select(ref_inp, dim, ref_index)
+    with flag_gems.use_gems():
+        res_out = torch.index_select(inp, dim, index)
+
+    utils.gems_assert_equal(res_out, ref_out)
