+++ tests/test_select_scatter.py
@@ -1,0 +1,56 @@
+import random
+import time
+
+import pytest
+import torch
+
+import flag_gems
+
+from . import accuracy_utils as utils
+from . import conftest as cfg
+
+if cfg.QUICK_MODE:
+    FLOAT_DTYPES = [torch.float32]
+    DIM_LIST = [1]
+else:
+    FLOAT_DTYPES = utils.FLOAT_DTYPES
+    DIM_LIST = [0, 1]
+
+random.seed(time.time() // 100)
+
+
+@pytest.mark.select_scatter
+@pytest.mark.parametrize("shape", utils.REDUCTION_SHAPES)
+@pytest.mark.parametrize("dim", DIM_LIST)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_select_scatter(shape, dim, dtype):
+    index = random.randint(0, shape[dim] - 1)
+    inp = torch.randn(shape, dtype=dtype, device=flag_gems.device)
+
+    src_shape = list(inp.shape)
+    del src_shape[dim]
+    src = torch.randn(src_shape, dtype=dtype, device=flag_gems.device)
+
+    ref_inp = utils.to_reference(inp)
+    ref_src = utils.to_reference(src)
+    ref_out = torch.select_scatter(ref_inp, dim=dim, index=index, src=ref_src)
+    with flag_gems.use_gems():
+        res_out = torch.select_scatter(inp, dim=dim, index=index, src=src)
+
+    utils.gems_assert_equal(res_out, ref_out)
+
+
+@pytest.mark.select_scatter
+def test_select_scatter_with_self_overlapping_input():
+    dim = 0
+    index = 1
+    inp = torch.randn((1, 4), device=flag_gems.device).broadcast_to((3, 4))
+    src = torch.randn((4,), device=flag_gems.device)
+
+    ref_inp = utils.to_reference(inp)
+    ref_src = utils.to_reference(src)
+    ref_out = torch.select_scatter(ref_inp, dim=dim, index=index, src=ref_src)
+    with flag_gems.use_gems():
+        res_out = torch.select_scatter(inp, dim=dim, index=index, src=src)
+
+    utils.gems_assert_equal(res_out, ref_out)
